requirements.txt
ultralytics
opencv-python
PyQt5
streamlit
pillow
pandas
